Add files via upload
This test was done with katalon we cane use this file with selenium too. 
We can screenshot each step if we want. 
On the nominal case, is the typical usecae. 
Like a sanity check, the user log, select a product, check out and log out. 
On the second test we just checkin the error message on the login page. 

Sincerly.